Move string evaluator to another package
