+++ src/main/kotlin/com/mituuz/fuzzier/components/TestBenchComponent.kt
@@ -38,7 +38,7 @@ import com.intellij.ui.components.JBTextArea
 import com.intellij.ui.table.JBTable
 import com.intellij.uiDesigner.core.GridConstraints
 import com.intellij.uiDesigner.core.GridLayoutManager
-import com.mituuz.fuzzier.StringEvaluator
+import com.mituuz.fuzzier.entities.StringEvaluator
 import com.mituuz.fuzzier.entities.FuzzyMatchContainer
 import com.mituuz.fuzzier.settings.FuzzierSettingsService
 import com.mituuz.fuzzier.util.FuzzierUtil
