+++ src/main/kotlin/com/mituuz/fuzzier/util/InitialViewHandler.java
@@ -1,0 +1,4 @@
+package com.mituuz.fuzzier.util;
+
+public class InitialViewHandler {
+}
